requirements.txt
peewee
riotwatcher
python-dotenv